+++ src/Data/Rpm/DbHeaderBlob.hs
@@ -1,0 +1,9 @@
+module Data.Rpm.DbHeaderBlob (
+  PkgInfo (..),
+  readPackageInfo,
+) where
+
+import Data.Rpm.DbHeaderBlob.Internal (
+  PkgInfo (..),
+  readPackageInfo,
+ )